[pax] Applied patch 2.6.30-test3
This is pax-linux-2.6.30-test3.patch

+++ arch/x86/xen/enlighten.c
@@ -454,7 +454,7 @@ static void xen_write_idt_entry(gate_desc *dt, int entrynum, const gate_desc *g)
 
 	preempt_disable();
 
-	start = __get_cpu_var(idt_desc).address;
+	start = (unsigned long)__get_cpu_var(idt_desc).address;
 	end = start + __get_cpu_var(idt_desc).size + 1;
 
 	xen_mc_flush();
